refactor: combine different inputs into one universal one
