--- src/components/auth-input/auth-input.module.scss
@@ -1,3 +1,0 @@
-.form-item {
-  width: 100%;
-}
