--- src/components/task-title-input/task-title-input.module.scss
@@ -1,4 +1,0 @@
-.input {
-  max-width: 500px;
-  width: 100%;
-}
