+++ src/pages/auth-pages/registration-page/registration-page.module.scss
@@ -13,3 +13,7 @@
   text-align: center;
   margin-bottom: 20px;
 }
+
+.form-item {
+  width: 100%;
+}
