+++ src/pages/todos-page/todos-page.module.scss
@@ -19,3 +19,8 @@
   margin-bottom: 20px;
   width: 100%;
 }
+
+.input {
+  max-width: 500px;
+  width: 100%;
+}
